JSON Spawner Logic Complete
All logic for the JSON Spawners is in place, I now just need to load them up from JSON files as well as recreate all the custom spawners in JSON format.

This is all untested so there may be problems but the code is pretty solid so hopefully there will be no major problems.

+++ src/main/java/com/lycanitesmobs/core/spawner/trigger/TickSpawnTrigger.java
@@ -16,7 +16,7 @@ public class TickSpawnTrigger extends SpawnTrigger {
 	}
 
 	/** Called every player tick. **/
-	public void onTick(EntityPlayer player, int ticks) {
+	public void onTick(EntityPlayer player, long ticks) {
 		// Tick Rate:
 		if(ticks == 0 || ticks % this.tickRate != 0) {
 			return;
